ExpedicionCopias/requirements.txt
pypdf>=3.0.0
msal>=1.24.0
holidays>=0.34
openpyxl>=3.1.0
requests>=2.31.0
azure-identity>=1.15.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
